Don't store server-side generated report summaries in the database. Fixes #1082.
